Change return type for a method + add comment for a method

+++ app/src/main/java/com/example/mealer_project/data/models/Chef.java
@@ -128,7 +128,7 @@ public class Chef extends User {
      * Get the number of ratings done for a chef
      * @return Integer representing number of ratings
      */
-    public double getNumOfRatings() { return numOfRatings; }
+    public int getNumOfRatings() { return numOfRatings; }
 
     /**
      * Set the chef's number of ratings
@@ -138,6 +138,10 @@ public class Chef extends User {
         this.numOfRatings = numOfRatings;
     }
 
+    /**
+     * Gets the chef's overall rating
+     * @return Double representing the rating
+     */
     public double getChefRating(){
         return chefRatingSum/numOfRatings;
     }
